nova conexão com o banco de dados e contrução de novas telas
