--- config/connectionFirebase.js
@@ -1,17 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyBWRkZ20ZKqZUSEGvakQ_uRDx8klF5m78o",
-  authDomain: "spis-project.firebaseapp.com",
-  projectId: "spis-project",
-  storageBucket: "spis-project.appspot.com",
-  messagingSenderId: "561134350228",
-  appId: "1:561134350228:web:fbfebe7f697ab03d216981"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
