+++ models/categoria.js
@@ -1,0 +1,5 @@
+export class Categoria{
+    constructor(nome) {
+        this.nome = nome;
+    }
+}
