+++ models/entrada.js
@@ -1,0 +1,8 @@
+export class Entrada{
+    constructor(dataEntrada, idPeca, idTipoEntrada, quantidade) {
+        this.dataEntrada = dataEntrada;
+        this.idPeca = idPeca;
+        this.idTipoEntrada = idTipoEntrada;
+        this.quantidade = quantidade;
+    }
+}
