+++ models/marca.js
@@ -1,0 +1,5 @@
+export class Marca{
+    constructor(nome) {
+        this.nome = nome;
+    }
+}
