+++ models/modelo.js
@@ -1,0 +1,6 @@
+export class Modelo{
+    constructor(nome, idPeca) {
+        this.nome = nome;
+        this.idPeca = idPeca;
+    }
+}
