+++ models/nome.js
@@ -1,0 +1,5 @@
+export class Nome{
+    constructor(nome) {
+        this.nome = nome;
+    }
+}
