+++ models/peca.js
@@ -1,0 +1,13 @@
+export class Peca{
+    constructor(descricao, idCategoria, idMarca, idNome, numeroSerie, quantidade, valorUnitarioCusto, valorUnitarioVenda) {
+        this.descricao = descricao;
+        this.idCategoria = idCategoria;
+        this.idMarca = idMarca;
+        this.idMarca = idMarca;
+        this.idNome = idNome;
+        this.numeroSerie = numeroSerie;
+        this.quantidade = quantidade;
+        this.valorUnitarioCusto = valorUnitarioCusto;
+        this.valorUnitarioVenda = valorUnitarioVenda;
+    }
+}
