+++ models/saida.js
@@ -1,0 +1,7 @@
+export class Entrada{
+    constructor(dataSaida, valorTotal, idPeca) {
+        this.dataSaida = dataSaida;
+        this.valorTotal = valorTotal;
+        this.idPeca = idPeca;
+    }
+}
