+++ models/tipoEntrada.js
@@ -1,0 +1,5 @@
+export class TipoEntrada{
+    constructor(nome) {
+        this.nome = nome;
+    }
+}
