+++ models/tipoSaida.js
@@ -1,0 +1,5 @@
+export class TipoSaida{
+    constructor(nome) {
+        this.nome = nome;
+    }
+}
