+++ package.json
@@ -9,14 +9,19 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-community/checkbox": "^0.5.14",
+    "@react-native-firebase/app": "^16.4.6",
     "@react-navigation/native": "^6.1.0",
+    "@react-navigation/native-stack": "^6.9.5",
     "cors": "^2.8.5",
     "expo": "~47.0.8",
+    "expo-dev-client": "~2.0.1",
     "expo-linear-gradient": "~12.0.1",
     "expo-status-bar": "~1.4.2",
-    "firebase": "^9.15.0",
     "react": "18.1.0",
-    "react-native": "0.70.5"
+    "react-native": "0.70.5",
+    "react-native-paper": "^5.0.1",
+    "react-native-safe-area-context": "^4.4.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
